requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
assemblyai
google-genai
jinja2
